Extract contact fucntion to contact module
